# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DurianFarm AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, done]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create an equipment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update an equipment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete an equipment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        category:
          type: ["string", "null"]
        plotId:
          type: ["integer", "null"]
        workerId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        category:
          type: string
        plotId:
          type: integer
        workerId:
          type: integer
        dueDate:
          type: string
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        category:
          type: string
        plotId:
          type: integer
        workerId:
          type: integer
        dueDate:
          type: string
          format: date
...
    InventoryItem:
      type: object
      required: [id, name, category, unit, quantity, minQuantity, costPerUnit, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minQuantity:
          type: number
        costPerUnit:
          type: number
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InventoryItemInput:
      type: object
      required: [name, category, unit]
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minQuantity:
          type: number
        costPerUnit:
          type: number
        supplier:
          type: string
        notes:
          type: string
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minQuantity:
          type: number
        costPerUnit:
          type: number
        supplier:
          type: string
        notes:
          type: string
...
    Equipment:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [operational, maintenance, broken, retired]
        purchaseDate:
          type: ["string", "null"]
          format: date
        lastMaintenanceDate:
          type: ["string", "null"]
          format: date
        nextMaintenanceDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EquipmentInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [operational, maintenance, broken, retired]
        purchaseDate:
          type: string
          format: date
        lastMaintenanceDate:
          type: string
          format: date
        nextMaintenanceDate:
          type: string
          format: date
        notes:
          type: string
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [operational, maintenance, broken, retired]
        purchaseDate:
          type: string
          format: date
        lastMaintenanceDate:
          type: string
          format: date
        nextMaintenanceDate:
          type: string
          format: date
        notes:
          type: string
    Notification:
      type: object
      required: [id, type, title, body, severity, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: string
        severity:
          type: string
          enum: [info, warning, urgent]
        read:
          type: boolean
        relatedPath:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...