# lib/api-zod/src/generated/types/equipmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DurianFarm AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EquipmentInputStatus } from './equipmentInputStatus';

export interface EquipmentInput {
  name: string;
  type: string;
  status?: EquipmentInputStatus;
  purchaseDate?: Date;
  lastMaintenanceDate?: Date;
  nextMaintenanceDate?: Date;
  notes?: string;
}