# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DurianFarm AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  name?: string;
  category?: string;
  unit?: string;
  quantity?: number;
  minQuantity?: number;
  costPerUnit?: number;
  supplier?: string;
  notes?: string;
}